+++ src/pydiverse/pipedag/backend/table/sql/dialects/mssql.py
@@ -15,6 +15,7 @@ from sqlalchemy import URL, Engine
 
 from pydiverse.common import Datetime, Dtype
 from pydiverse.pipedag.backend.table.sql.ddl import (
+    AddClusteredColumnstoreIndex,
     ChangeColumnTypes,
     CreateAlias,
     _mssql_update_definition,
@@ -27,11 +28,36 @@ from pydiverse.pipedag.backend.table.sql.hooks import (
 from pydiverse.pipedag.backend.table.sql.reflection import PipedagMSSqlReflection
 from pydiverse.pipedag.backend.table.sql.sql import SQLTableStore
 from pydiverse.pipedag.container import RawSql, Schema, Table
+from pydiverse.pipedag.materialize.details import (
+    BaseMaterializationDetails,
+    resolve_materialization_details_label,
+)
+
+
+@dataclass(frozen=True)
+class MSSqlMaterializationDetails(BaseMaterializationDetails):
+    """
+    Materialization details specific to Microsoft SQL Server.
+
+    :param columnstore: Whether to create tables as columnstores
+        by defining a clustered columnstore index on them. This can
+        improve performance.
+        See the `Microsoft documentation <https://learn.microsoft.com/en-us/sql/relational-databases/indexes/columnstore-indexes-overview>`_.
+    """
+
+    def __post_init__(self):
+        assert isinstance(self.columnstore, bool)
+
+    columnstore: bool = False
 
 
 class MSSqlTableStore(SQLTableStore):
     """
     SQLTableStore that supports `Microsoft SQL Server`_.
+
+    Supports the :py:class:`MSSqlMaterializationDetails\
+    <pydiverse.pipedag.backend.table.sql.dialects.mssql.MSSqlMaterializationDetails>`
+    materialization details.
 
     In addition to the arguments of
     :py:class:`SQLTableStore <pydiverse.pipedag.backend.table.SQLTableStore>`,
@@ -92,6 +118,7 @@ class MSSqlTableStore(SQLTableStore):
             or table.non_nullable is not None
             or (table.primary_key is not None and len(table.primary_key) > 0)
             or (table.indexes is not None and len(table.indexes) > 0)
+            or self.get_create_columnstore_index(table)
         )
 
     def get_forced_nullability_columns(
@@ -189,6 +216,8 @@ class MSSqlTableStore(SQLTableStore):
                             cap_varchar_max=1024,
                         )
                     )
+            if self.get_create_columnstore_index(table):
+                self.execute(AddClusteredColumnstoreIndex(table.name, schema))
         if on_empty_table is None or not on_empty_table:
             self.add_table_primary_key(table, schema)
             self.add_table_indexes(table, schema)
@@ -294,6 +323,28 @@ class MSSqlTableStore(SQLTableStore):
         # For normal Table objects, it needs the stage schema name.
         table_name, schema = super().resolve_alias(table, stage_name)
         return PipedagMSSqlReflection.resolve_alias(self.engine, table_name, schema)
+
+    def _set_materialization_details(
+        self, materialization_details: dict[str, dict[str | list[str]]] | None
+    ) -> None:
+        self.materialization_details = (
+            MSSqlMaterializationDetails.create_materialization_details_dict(
+                materialization_details,
+                self.strict_materialization_details,
+                self.default_materialization_details,
+                self.logger,
+            )
+        )
+
+    def get_create_columnstore_index(self, table: Table) -> bool:
+        return MSSqlMaterializationDetails.get_attribute_from_dict(
+            self.materialization_details,
+            resolve_materialization_details_label(table),
+            self.default_materialization_details,
+            "columnstore",
+            self.strict_materialization_details,
+            self.logger,
+        )
 
 
 @MSSqlTableStore.register_table(pd)