+++ docs/source/changelog.md
@@ -8,6 +8,12 @@
   and extra year columns any more
 - Materialize Local Table Cache before actual Table Store
 
+## 0.9.11 (2025-XX-XX)
+- Fix numpy import issue on OS X
+- CI: Use SQL Server docker image instead of Azure Edge
+- Improve documentation for materialization details.
+- Add support for columnstore tables in `MSSQLTableStore` via `MSSqlMaterializationDetails`
+
 ## 0.9.10 (2025-03-18)
 - Fix incompatibility with pydiverse.transform >=0.2.1 (<0.2.0 still supported) (0.2.0 will not be supported):
   - pydiverse transform had a radical refactoring around this point